Refactored code to be able to convert the request result to the new model which is a record with several components, as opposed to just the weather data.
This does not compile, but all problems are refactored into a single conflict point, which needs to be resolved. To resolve it i need to figure out how Commands work properly.

Alternatively it might very well be the correct approach to NOT solve this issue this way, and instead attack it from a completely different angle, and not running this request immediately upon initalization, and applying it to the model, and instead doing this as just one part of the data flow. For instance having a different start state which is just idle or similar.